refactor: ParticipantRepository 제거
우선 Participant를 Club과 하나의 Aggregate로 묶는다.
