+++ mohaeng/src/main/java/com/mohaeng/club/applicationform/application/service/ApproveApplicationForm.java
@@ -6,28 +6,23 @@ import com.mohaeng.club.applicationform.domain.event.OfficerApproveApplicationEv
 import com.mohaeng.club.applicationform.domain.model.ApplicationForm;
 import com.mohaeng.club.applicationform.domain.repository.ApplicationFormRepository;
 import com.mohaeng.club.applicationform.exception.ApplicationFormException;
-import com.mohaeng.club.participant.domain.model.Participant;
-import com.mohaeng.club.participant.domain.repository.ParticipantRepository;
-import com.mohaeng.club.participant.exception.ParticipantException;
+import com.mohaeng.club.club.domain.model.Participant;
+import com.mohaeng.club.club.exception.ParticipantException;
 import com.mohaeng.common.event.Events;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import static com.mohaeng.club.applicationform.exception.ApplicationFormExceptionType.NOT_FOUND_APPLICATION_FORM;
-import static com.mohaeng.club.participant.exception.ParticipantExceptionType.NOT_FOUND_PARTICIPANT;
-import static com.mohaeng.club.participant.exception.ParticipantExceptionType.NOT_FOUND_PRESIDENT;
+import static com.mohaeng.club.club.exception.ParticipantExceptionType.NOT_FOUND_PARTICIPANT;
 
 @Service
 @Transactional
 public class ApproveApplicationForm implements ApproveApplicationFormUseCase {
 
     private final ApplicationFormRepository applicationFormRepository;
-    private final ParticipantRepository participantRepository;
 
-    public ApproveApplicationForm(final ApplicationFormRepository applicationFormRepository, final ParticipantRepository participantRepository) {
+    public ApproveApplicationForm(final ApplicationFormRepository applicationFormRepository) {
         this.applicationFormRepository = applicationFormRepository;
-        this.participantRepository = participantRepository;
-
     }
 
     @Override
@@ -35,7 +30,7 @@ public class ApproveApplicationForm implements ApproveApplicationFormUseCase {
         ApplicationForm applicationForm = applicationFormRepository.findById(command.applicationFormId())
                 .orElseThrow(() -> new ApplicationFormException(NOT_FOUND_APPLICATION_FORM));
 
-        Participant manager = participantRepository.findByMemberIdAndClubId(command.managerId(), applicationForm.club().id())
+        Participant manager = applicationForm.club().findParticipantByMemberId(command.managerId())
                 .orElseThrow(() -> new ParticipantException(NOT_FOUND_PARTICIPANT));
 
         applicationForm.approve(manager);
@@ -51,12 +46,10 @@ public class ApproveApplicationForm implements ApproveApplicationFormUseCase {
                 applicationForm.club().id()));
 
         // 2. 회장이 아닌 경우, 회장에게 누가 가입 신청서를 승인했다는 알림
-        Participant president = participantRepository.findPresidentByClubId(applicationForm.club().id())
-                .orElseThrow(() -> new ParticipantException(NOT_FOUND_PRESIDENT));
-        if (president.id().equals(manager.id())) {
+        if (manager.isPresident()) {
             return;
         }
-
+        Participant president = applicationForm.club().findPresident();
         Long applicantId = applicationForm.applicant().id();
         Participant registerParticipant = applicationForm.club().findParticipantByMemberId(applicantId).orElseThrow(IllegalStateException::new);
         Events.raise(new OfficerApproveApplicationEvent(this,
