+++ mohaeng/src/main/java/com/mohaeng/club/applicationform/application/service/FillOutApplicationForm.java
@@ -6,9 +6,9 @@ import com.mohaeng.club.applicationform.domain.model.ApplicationForm;
 import com.mohaeng.club.applicationform.domain.repository.ApplicationFormRepository;
 import com.mohaeng.club.applicationform.exception.ApplicationFormException;
 import com.mohaeng.club.club.domain.model.Club;
+import com.mohaeng.club.club.domain.model.Participant;
 import com.mohaeng.club.club.domain.repository.ClubRepository;
 import com.mohaeng.club.club.exception.ClubException;
-import com.mohaeng.club.participant.domain.model.Participant;
 import com.mohaeng.common.domain.BaseEntity;
 import com.mohaeng.common.event.Events;
 import com.mohaeng.member.domain.model.Member;
