+++ mohaeng/src/test/java/com/mohaeng/club/applicationform/application/service/FillOutApplicationFormTest.java
@@ -6,9 +6,8 @@ import com.mohaeng.club.applicationform.domain.model.ApplicationForm;
 import com.mohaeng.club.applicationform.domain.repository.ApplicationFormRepository;
 import com.mohaeng.club.applicationform.exception.ApplicationFormException;
 import com.mohaeng.club.club.domain.model.Club;
+import com.mohaeng.club.club.domain.model.Participant;
 import com.mohaeng.club.club.domain.repository.ClubRepository;
-import com.mohaeng.club.participant.domain.model.Participant;
-import com.mohaeng.club.participant.domain.repository.ParticipantRepository;
 import com.mohaeng.common.annotation.ApplicationTest;
 import com.mohaeng.common.exception.BaseExceptionType;
 import com.mohaeng.member.domain.model.Member;
@@ -48,9 +47,6 @@ class FillOutApplicationFormTest {
 
     @Autowired
     private ClubRepository clubRepository;
-
-    @Autowired
-    private ParticipantRepository participantRepository;
 
     @Autowired
     private ApplicationEvents events;
