+++ mohaeng/src/test/java/com/mohaeng/club/applicationform/domain/ApplicationFormTest.java
@@ -3,9 +3,9 @@ package com.mohaeng.club.applicationform.domain;
 import com.mohaeng.club.applicationform.domain.model.ApplicationForm;
 import com.mohaeng.club.applicationform.exception.ApplicationFormException;
 import com.mohaeng.club.club.domain.model.Club;
+import com.mohaeng.club.club.domain.model.Participant;
 import com.mohaeng.club.club.exception.ClubException;
-import com.mohaeng.club.participant.domain.model.Participant;
-import com.mohaeng.club.participant.exception.ParticipantException;
+import com.mohaeng.club.club.exception.ParticipantException;
 import com.mohaeng.common.exception.BaseExceptionType;
 import com.mohaeng.common.fixtures.ParticipantFixture;
 import com.mohaeng.member.domain.model.Member;
@@ -14,7 +14,7 @@ import org.junit.jupiter.api.*;
 import static com.mohaeng.club.applicationform.exception.ApplicationFormExceptionType.ALREADY_PROCESSED;
 import static com.mohaeng.club.applicationform.exception.ApplicationFormExceptionType.NO_AUTHORITY_PROCESS_APPLICATION;
 import static com.mohaeng.club.club.exception.ClubExceptionType.CLUB_IS_FULL;
-import static com.mohaeng.club.participant.exception.ParticipantExceptionType.ALREADY_EXIST_PARTICIPANT;
+import static com.mohaeng.club.club.exception.ParticipantExceptionType.ALREADY_EXIST_PARTICIPANT;
 import static com.mohaeng.common.fixtures.ApplicationFormFixture.applicationForm;
 import static com.mohaeng.common.fixtures.ClubFixture.FULL_CLUB;
 import static com.mohaeng.common.fixtures.ClubFixture.club;
